chore: fixed flaky tests
