+++ Scripts/test-carthage-integration.sh
@@ -116,6 +116,7 @@ schemes:
 mkdir -p Tests
 mkdir -p Source
 cp "../Tests/AdyenTests/Adyen Tests/DropIn/DropInTests.swift" Tests/DropInTests.swift
+cp "../Tests/AdyenTests/Adyen Tests/DropIn/XCTestCaseExtensions.swift" Tests/XCTestCaseExtensions.swift
 cp "../Tests/AdyenTests/Adyen Tests/Components/Dummy.swift" Tests/Dummy.swift
 cp -a "../Demo/Common" Source/
 cp -a "../Demo/UIKit" Source/
