+++ Tests/AdyenTests/Adyen Tests/Core/BackoffSchedulerTests.swift
@@ -10,18 +10,6 @@
 @testable import AdyenNetworking
 import XCTest
 
-extension XCTestCase {
-    func wait(for interval: DispatchTimeInterval) {
-        let dummyExpectation = XCTestExpectation(description: "wait for a few seconds.")
-
-        DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + interval) {
-            dummyExpectation.fulfill()
-        }
-
-        wait(for: [dummyExpectation], timeout: 100)
-    }
-}
-
 class BackoffSchedulerTests: XCTestCase {
 
     func testScheduling() {
